requirements.txt
python-dotenv
requests
aiohttp
tenacity
pyodbc